peer review refactorings
