+++ src/modules/tests/test-data/cat-d/test-data.cat-d.selector.ts
@@ -42,7 +42,7 @@ export const getDrivingFaultCount = (
 export const getManoeuvres = (data: CatDTestData): CatDManoeuvres => data.manoeuvres;
 
 // TODO - We should pass a Manoeuvre object here instead of TestData
-export const hasManoeuvreBeenCompletedCatC = (data: CatDTestData) => {
+export const hasManoeuvreBeenCompletedCatD = (data: CatDTestData) => {
   return (
     get(data.manoeuvres, 'reverseLeft.selected')
   );
