+++ src/pages/test-report/__tests__/test-report.analytics.effects.spec.ts
@@ -12,7 +12,7 @@ import * as seriousFaultsActions from '../../../modules/tests/test-data/common/s
 import * as manoeuvresActions from '../../../modules/tests/test-data/common/manoeuvres/manoeuvres.actions';
 import * as vehicleChecksActions from '../../../modules/tests/test-data/cat-b/vehicle-checks/vehicle-checks.actions';
 import * as testRequirementsActions
-  from '../../../modules/tests/test-data/cat-b/test-requirements/test-requirements.actions';
+  from '../../../modules/tests/test-data/common/test-requirements/test-requirements.actions';
 import * as ecoActions from '../../../modules/tests/test-data/common/eco/eco.actions';
 import * as testReportActions from '../test-report.actions';
 import { StoreModel } from '../../../shared/models/store.model';
