+++ src/pages/test-report/components/legal-requirement/__tests__/legal-requirement.spec.ts
@@ -9,7 +9,7 @@ import { StoreModule, Store } from '@ngrx/store';
 import { StoreModel } from '../../../../../shared/models/store.model';
 import {
   ToggleLegalRequirement,
-} from '../../../../../modules/tests/test-data/cat-b/test-requirements/test-requirements.actions';
+} from '../../../../../modules/tests/test-data/common/test-requirements/test-requirements.actions';
 import { LegalRequirements } from '../../../../../modules/tests/test-data/test-data.constants';
 import { NavigationStateProvider } from '../../../../../providers/navigation-state/navigation-state';
 import { NavigationStateProviderMock } from '../../../../../providers/navigation-state/__mocks__/navigation-state.mock';
